Implement aio best practice on main gateway app
