+++ __main__.py
@@ -1,4 +1,1 @@
-from src.gateway import *
-
-gw = Gateway()
-asyncio.run(gw.main_loop())
+from src.app import *
